+++ click_extended/decorators/check/__init__.py
@@ -4,18 +4,22 @@ from click_extended.decorators.check.conflicts import conflicts
 from click_extended.decorators.check.dependencies import dependencies
 from click_extended.decorators.check.ends_with import ends_with
 from click_extended.decorators.check.exclusive import exclusive
+from click_extended.decorators.check.falsy import falsy
 from click_extended.decorators.check.length import length
 from click_extended.decorators.check.not_empty import not_empty
 from click_extended.decorators.check.requires import requires
 from click_extended.decorators.check.starts_with import starts_with
+from click_extended.decorators.check.truthy import truthy
 
 __all__ = [
     "conflicts",
     "dependencies",
     "ends_with",
     "exclusive",
+    "falsy",
     "length",
     "not_empty",
     "requires",
     "starts_with",
+    "truthy",
 ]